Mark step labels for translation and de-dupe string helper functions
