+++ frontend/awx/resources/templates/TemplatePage/steps/ConditionalField.tsx
@@ -1,0 +1,9 @@
+export function ConditionalField({
+  isHidden = false,
+  children,
+}: {
+  isHidden: boolean;
+  children: React.ReactNode;
+}) {
+  return isHidden ? null : <>{children}</>;
+}
